Introduce a new AppM module, under IO.
@noteed the reason why I've put this under IO since it provides
functionality over IO.
