+++ src/Curiosity/Runtime/IO/AppM.hs
@@ -1,0 +1,131 @@
+{-# LANGUAGE TypeFamilies  #-}
+{-# LANGUAGE DeriveFunctor #-}
+{-# LANGUAGE GeneralisedNewtypeDeriving #-}
+module Curiosity.Runtime.IO.AppM
+  ( AppM(..)
+  , runAppMSafe
+  , checkCredentials
+  ) where
+
+import qualified Commence.Multilogging         as ML
+import qualified Commence.Runtime.Errors       as Errs
+import qualified Commence.Runtime.Storage      as S
+import           Commence.Types.Secret          ( (=:=) )
+import qualified Control.Concurrent.STM        as STM
+import           Control.Lens                  as Lens
+import "exceptions" Control.Monad.Catch         ( MonadCatch
+                                                , MonadMask
+                                                , MonadThrow
+                                                )
+import qualified Curiosity.Core                as Core
+import qualified Curiosity.Data                as Data
+import qualified Curiosity.Data.User           as User
+import           Curiosity.Runtime.Type        as RType
+import           Prelude                 hiding ( state )
+
+
+--------------------------------------------------------------------------------
+newtype AppM a = AppM { runAppM :: ReaderT Runtime (ExceptT Errs.RuntimeErr IO) a }
+  deriving ( Functor
+           , Applicative
+           , Monad
+           , MonadIO
+           , MonadReader Runtime
+           , MonadError Errs.RuntimeErr
+           , MonadThrow
+           , MonadCatch
+           , MonadMask
+           )
+
+-- | Run the `AppM` computation catching all possible exceptions.
+runAppMSafe
+  :: forall a m
+   . MonadIO m
+  => Runtime
+  -> AppM a
+  -> m (Either Errs.RuntimeErr a)
+runAppMSafe runtime AppM {..} =
+  liftIO
+    . fmap (join . first Errs.RuntimeException)
+    . try @SomeException
+    . runExceptT
+    $ runReaderT runAppM runtime
+
+
+--------------------------------------------------------------------------------
+-- | Support for logging for the application
+instance ML.MonadAppNameLogMulti AppM where
+  askLoggers = asks _rLoggers
+  localLoggers modLogger =
+    local (over rLoggers . over ML.appNameLoggers $ fmap modLogger)
+
+instance S.DBTransaction AppM STM where
+  liftTxn =
+    liftIO
+      . fmap (first Errs.RuntimeException)
+      . try @SomeException
+      . STM.atomically
+
+
+--------------------------------------------------------------------------------
+-- | Definition of all operations for the UserProfiles (selects and updates)
+instance S.DBStorage AppM STM User.UserProfile where
+
+  type Db AppM STM User.UserProfile = Core.StmDb
+
+  type DBError AppM STM User.UserProfile = User.Err
+
+  dbUpdate db@Data.Db {..} = \case
+
+    User.UserCreate input -> second pure <$> Core.createUserFull db input
+
+    User.UserCreateGeneratingUserId input ->
+      second (pure . fst) <$> Core.createUser db input
+
+    User.UserDelete id ->
+      S.dbSelect @AppM @STM db (User.SelectUserById id) <&> headMay >>= maybe
+        (pure . userNotFound $ show id)
+        (fmap Right . deleteUser)
+     where
+      deleteUser _ =
+        modifyUserProfiles id (filter $ (/= id) . S.dbId) _dbUserProfiles
+
+    User.UserUpdate id (User.Update mname mbio) ->
+      S.dbSelect @AppM @STM db (User.SelectUserById id) <&> headMay >>= maybe
+        (pure . userNotFound $ show id)
+        (fmap Right . updateUser)
+     where
+      updateUser _ = modifyUserProfiles id replaceOlder _dbUserProfiles
+      change =
+        set User.userProfileBio mbio . set User.userProfileDisplayName mname
+      replaceOlder users =
+        [ if S.dbId u == id then change u else u | u <- users ]
+
+  dbSelect db = \case
+
+    User.UserLoginWithUserName input -> toList <$> checkCredentials db input
+
+    User.SelectUserById        id    -> toList <$> Core.selectUserById db id
+
+    User.SelectUserByUserName username ->
+      toList <$> Core.selectUserByUsername db username
+
+userNotFound = Left . User.UserNotFound . mappend "User not found: "
+
+modifyUserProfiles id f userProfiles = STM.modifyTVar userProfiles f $> [id]
+
+checkCredentials
+  :: Core.StmDb -> User.Credentials -> STM (Maybe User.UserProfile)
+checkCredentials db User.Credentials {..} = do
+  mprofile <- Core.selectUserByUsername db _userCredsName
+  case mprofile of
+    Just profile | checkPassword profile _userCredsPassword ->
+      pure $ Just profile
+    _ -> pure Nothing
+
+-- TODO Use constant-time string comparison.
+checkPassword :: User.UserProfile -> User.Password -> Bool
+checkPassword profile (User.Password passInput) = storedPass =:= passInput
+ where
+  User.Password storedPass =
+    profile ^. User.userProfileCreds . User.userCredsPassword
